+++ R/calc_sdp.R
@@ -11,7 +11,7 @@
 #' 1 and \eqn{2^n - 2} where \eqn{n} is the number of strains, whose
 #' binary representation indicates the strain genotypes.
 #'
-#' @seealso [invert_sdp()]
+#' @seealso [invert_sdp()], [sdp2char()]
 #'
 #' @export
 #' @examples
@@ -65,7 +65,7 @@ calc_sdp <-
 #' (AA) and 3 (BB). Markers with values other than 1 or 3 are omitted,
 #' and monomorphic markers, are omitted.
 #'
-#' @seealso [calc_sdp()]
+#' @seealso [sdp2char()], [calc_sdp()]
 #'
 #' @export
 #' @examples
@@ -82,4 +82,48 @@ invert_sdp <-
     geno <- geno*2+1 # convert 0/1 to 1/3
 
     geno
+}
+
+#' Convert strain distribution patterns to character strings
+#'
+#' Convert a vector of numeric codes for strain distribution patterns to character strings.
+#'
+#' @param sdp Vector of strain distribution patterns (integers between
+#' 1 and \eqn{2^n-2} where \eqn{n} is the number of strains.
+#' @param n_strains
+#' @param strains Vector of single-letter codes for the strains
+#'
+#' @return Vector of character strings with the two groups of alleles
+#'     separated by a vertical bar (|).
+#'
+#' @seealso [invert_sdp()], [calc_sdp()]
+#'
+#' @export
+#' @examples
+#' sdp <- c(m1=1, m2=12, m3=240)
+#' sdp2char(sdp, LETTERS[1:8])
+sdp2char <-
+    function(sdp, n_strains=NULL, strains=NULL)
+{
+    if(is.null(n_strains) && !is.null(strains)) {
+        n_strains <- length(strains)
+    }
+    if(!is.null(n_strains) && is.null(strains)) {
+        strains <- LETTERS[seq_len(n_strains)]
+    }
+    if(is.null(n_strains) && is.null(strains)) {
+        stop("Provide at least one of n_strains or strains")
+    }
+
+    # convert to matrix
+    pat <- invert_sdp(sdp, n_strains)
+
+    # convert to character strings
+    result <- apply(pat, 1, function(p)
+        paste(paste(strains[p==3], collapse=""),
+              paste(strains[p==1], collapse=""), sep="|") )
+
+    names(result) <- names(sdp)
+
+    result
 }